Add files via upload
First versions of the catalogue files for the core rules companies plus the second version of the game system file. The game system file has some of the content from the first supplement (Heroes, Villains and Fiends) included but hidden until the work is complete. Also working on a 'build-your-own' company function as per the core rules, but this may take some time as it is really tedious!